refactor: Move remaining test
